Creating new visual for Color Tile
